Abstract type definitions to colocated file
