+++ packages/skeleton-react/src/components/Accordion/Accordion.tsx
@@ -1,14 +1,15 @@
 "use client";
 
-import React, {
-  createContext,
-  useContext,
-  useState,
-  useEffect,
-  type Dispatch,
-  type SetStateAction,
-  type ReactNode,
-} from "react";
+import React, { createContext, useContext, useState, useEffect } from "react";
+import {
+  AccordionContextState,
+  AccordionControlProps,
+  AccordionItemProps,
+  AccordionPanelProps,
+  AccordionProps,
+} from "./types";
+
+// React Compose ---
 
 // FIXME:
 // import { reactCompose } from "../../utils/react-compose";
@@ -19,78 +20,7 @@ export function reactCompose<
   return Object.assign(root, components);
 }
 
-interface AccordionProps extends React.PropsWithChildren {
-  /** Enables opening multiple items at once. */
-  multiple?: boolean;
-  /** The slide animation duration in milliseconds. */
-  // animDuration?: number;
-  // ---
-  /** Sets base styles. */
-  base?: string;
-  /** Set padding styles. */
-  padding?: string;
-  /** Set vertical spacing styles. */
-  spaceY?: string;
-  /** Set border radius styles. */
-  rounded?: string;
-  /** Set the width styles. */
-  width?: string;
-  /** Provide arbitrary CSS classes. */
-  classes?: string;
-}
-
-interface AccordionItemProps extends React.PropsWithChildren {
-  /** Sets base styles. */
-  base?: string;
-  /** Set vertical spacing styles. */
-  spaceY?: string;
-  /** Provide arbitrary CSS classes. */
-  classes?: string;
-}
-
-interface AccordionControlProps extends React.PropsWithChildren {
-  /** Identifies the panel ID this controls. */
-  controls: string;
-  /** Set the open state of the item. */
-  open?: boolean;
-  /** Set a disabled state for the item. */
-  disabled?: boolean;
-  // Root
-  /** Sets control's base styles. */
-  base?: string;
-  /** Sets control's the hover styles. */
-  hover?: string;
-  /** Sets control's the padding styles. */
-  padding?: string;
-  /** Sets control's the border radius styles. */
-  rounded?: string;
-  /** Provide arbitrary CSS classes to the control. */
-  classes?: string;
-  // Slots
-  lead?: ReactNode;
-}
-
-interface AccordionPanelProps extends React.PropsWithChildren {
-  /** The unique panel ID. */
-  id: string;
-  /** Set the panel's base styles. */
-  base?: string;
-  /** Set the panel's padding styles. */
-  padding?: string;
-  /** Set the panel's border-radius styles. */
-  rounded?: string;
-  /** Provide arbitrary CSS classes to the panel. */
-  classes?: string;
-}
-
 // Context ---
-
-interface AccordionContextState {
-  selected: string[];
-  setSelected: Dispatch<SetStateAction<string[]>>;
-  allowMultiple: boolean;
-  setAllowMultiple: Dispatch<SetStateAction<boolean>>;
-}
 
 const AccordionContext = createContext<AccordionContextState>({
   selected: [],
@@ -99,7 +29,8 @@ const AccordionContext = createContext<AccordionContextState>({
   setAllowMultiple: () => {},
 });
 
-/** Component: An Accordion child element. */
+// Components ---
+
 const AccordionRoot: React.FC<AccordionProps> = ({
   multiple = false,
   // Root
@@ -133,7 +64,6 @@ const AccordionRoot: React.FC<AccordionProps> = ({
   );
 };
 
-/** Component: An Accordion child element. */
 const AccordionItem: React.FC<AccordionItemProps> = ({
   base = "",
   spaceY = "",
