+++ packages/skeleton-svelte/src/lib/components/Accordion/AccordionItem.svelte
@@ -2,6 +2,8 @@
 	import { getContext, untrack, type Snippet } from 'svelte';
 	import { slide } from 'svelte/transition';
 
+	import type { AccordionItemProps } from './types.js';
+
 	// FIXME:
 	// import { State } from '$lib/utils.svelte.js';
 	export class State<T> {
@@ -9,62 +11,6 @@
 		constructor(initial: T) {
 			this.value = initial;
 		}
-	}
-
-	type ToggleEvent = CustomEvent<{ id: string; open: boolean }>;
-
-	interface AccordionItemProps {
-		/** Set a unique ID for the item. */
-		id: string;
-		/** Set the open state of the item. */
-		open?: boolean;
-		/** Set a disabled state for the item. */
-		disabled?: boolean;
-
-		// Root ---
-		/** Sets base styles. */
-		base?: string;
-		/** Set vertical spacing styles. */
-		spaceY?: string;
-		/** Provide arbitrary CSS classes. */
-		classes?: string;
-
-		// Control ---
-		/** Sets control's base styles. */
-		controlBase?: string;
-		/** Sets control's the hover styles. */
-		controlHover?: string;
-		/** Sets control's the padding styles. */
-		controlPadding?: string;
-		/** Sets control's the border radius styles. */
-		controlRounded?: string;
-		/** Provide arbitrary CSS classes to the control. */
-		controlClasses?: string;
-
-		// Icons ---
-		/** Set the base styles for the state icons. */
-		iconsBase?: string;
-
-		// Panel ---
-		/** Set the panel's base styles. */
-		panelBase?: string;
-		/** Set the panel's padding styles. */
-		panelPadding?: string;
-		/** Set the panel's border-radius styles. */
-		panelRounded?: string;
-		/** Provide arbitrary CSS classes to the panel. */
-		panelClasses?: string;
-
-		// Events ---
-		/** Handle the toggle event. */
-		ontoggle?: (toggleEvent: ToggleEvent) => void;
-		// Snippets ---
-		/** The control's default slot. */
-		control: Snippet;
-		/** The control's lead icon slot. */
-		controlLead?: Snippet;
-		/** The panels's default slot. */
-		panel?: Snippet;
 	}
 
 	let {
